ditch the mmio module
