+++ examples/mode4.rs
@@ -3,9 +3,12 @@
 
 use gba::{
   bios::VBlankIntrWait,
-  mmio::{BG_PALRAM, DISPCNT, DISPSTAT, IE, IME, KEYINPUT, MODE4_VRAM},
-  video::{Color, DisplayControl, DisplayStatus},
-  IrqBits,
+  irq::{IrqBits, IE, IME},
+  keys::KEYINPUT,
+  video::{
+    Color, DisplayControl, DisplayStatus, BG_PALRAM, DISPCNT, DISPSTAT,
+    MODE4_VRAM,
+  },
 };
 
 gba::panic_handler!(empty_loop);
